Change VisualComponent struct to suit in memory struct (output); list(z->list(x,y))
