+++ src/main/java/edu/uiuc/zenvisage/data/remotedb/VisualComponentList.java
@@ -1,0 +1,15 @@
+package edu.uiuc.zenvisage.data.remotedb;
+
+import java.util.ArrayList;
+
+public class VisualComponentList {
+	private ArrayList<VisualComponent> visualComponent;
+
+	public ArrayList<VisualComponent> getVisualComponent() {
+		return visualComponent;
+	}
+
+	public void setVisualComponent(ArrayList<VisualComponent> visualComponent) {
+		this.visualComponent = visualComponent;
+	}
+}
